Add HearthstoneJSON (#101)
There is a cards.json, cards.collectible.json and cardbacks.json for every game patch to date. By using "latest", it automatically redirects to the latest patch.

### cards.json
Every card in the game, and since everything in Hearthstone is a card, this includes non-playable cards such as those created by spells
### cards.collectible.json
Cards that can be collected, as well as heroes.
### cardbacks.json
Details of all the available card backs.

Each JSON file is available either in English or in every available language, but the latter is quite large, so I only included one example.